+++ src/EventSubscriber/GetAvailableResourceClassActionsEventSubscriber.php
@@ -10,10 +10,58 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 class GetAvailableResourceClassActionsEventSubscriber implements EventSubscriberInterface
 {
-    private const AVAILABLE_FORM_ITEM_ACTIONS = AuthorizationService::FORM_ITEM_ACTIONS;
-    private const AVAILABLE_FORM_COLLECTION_ACTIONS = [AuthorizationService::CREATE_FORMS_ACTION];
+    private const AVAILABLE_FORM_ITEM_ACTIONS = [
+        AuthorizationService::READ_FORM_ACTION => [
+            'en' => 'Read',
+            'de' => 'Lesen',
+        ],
+        AuthorizationService::UPDATE_FORM_ACTION => [
+            'en' => 'Update',
+            'de' => 'Aktualisieren',
+        ],
+        AuthorizationService::DELETE_FORM_ACTION => [
+            'en' => 'Delete',
+            'de' => 'Löschen',
+        ],
+        AuthorizationService::CREATE_SUBMISSIONS_FORM_ACTION => [
+            'en' => 'Create Submissions',
+            'de' => 'Einreichungen erstellen',
+        ],
+        AuthorizationService::READ_SUBMISSIONS_FORM_ACTION => [
+            'en' => 'Read Submissions',
+            'de' => 'Einreichungen lesen',
+        ],
+        AuthorizationService::UPDATE_SUBMISSIONS_FORM_ACTION => [
+            'en' => 'Update Submissions',
+            'de' => 'Einreichungen aktualisieren',
+        ],
+        AuthorizationService::DELETE_SUBMISSIONS_FORM_ACTION => [
+            'en' => 'Delete Submissions',
+            'de' => 'Einreichungen löschen',
+        ],
+    ];
 
-    private const AVAILABLE_SUBMISSION_ITEM_ACTIONS = AuthorizationService::SUBMISSION_ITEM_ACTIONS;
+    private const AVAILABLE_FORM_COLLECTION_ACTIONS = [
+        AuthorizationService::CREATE_FORMS_ACTION => [
+            'en' => 'Create',
+            'de' => 'Erstellen',
+        ],
+    ];
+
+    private const AVAILABLE_SUBMISSION_ITEM_ACTIONS = [
+        AuthorizationService::READ_SUBMISSION_ACTION => [
+            'en' => 'Read',
+            'de' => 'Lesen',
+        ],
+        AuthorizationService::UPDATE_SUBMISSION_ACTION => [
+            'en' => 'Update',
+            'de' => 'Aktualisieren',
+        ],
+        AuthorizationService::DELETE_SUBMISSION_ACTION => [
+            'en' => 'Delete',
+            'de' => 'Löschen',
+        ],
+    ];
     private const AVAILABLE_SUBMISSION_COLLECTION_ACTIONS = [];
 
     public static function getSubscribedEvents(): array